+++ .travis.yml
@@ -1,0 +1,17 @@
+language: go
+
+go:
+  - 1.5
+  - 1.6
+  - 1.7
+  - 1.7.1
+
+before_install:
+  - go get github.com/tools/godep
+  - GO15VENDOREXPERIMENT=1 godep restore -v
+
+script:
+- GO15VENDOREXPERIMENT=1 godep go test -race -v github.com/goldeneggg/lsec2
+- GO15VENDOREXPERIMENT=1 godep go test -race -v github.com/goldeneggg/lsec2/awsec2
+
+sudo: false
