+++ constants/const.go
@@ -4,3 +4,11 @@ const (
 	ExitStsOk = iota
 	ExitStsNg
 )
+
+const (
+	TagFilterPrefix   = "tag:"
+	TagPairSeparator  = "="
+	TagValueSeparator = ","
+
+	UndefinedItem = "UNDEFINED"
+)
